# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "\nRosetta translates vehicle telemetry from many car makers into one canonical event.\n\n* **Authentication** OAuth2 bearer tokens (JWT). Get one from `POST /api/v1/auth/token`.\n* **Errors** RFC 9457 problem details (`application/problem+json`).\n* **Pagination** keyset cursors: pass `next_cursor` back as `cursor`. No offsets.\n* **Rate limits** per caller. A 429 carries `Retry-After`.\n* **Versioning** the path carries the major version. Additive changes do not bump it.\n",
    "title": "Rosetta API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ingest/{oem}": {
      "post": {
        "operationId": "ingest_api_v1_ingest__oem__post",
        "parameters": [
          {
            "in": "path",
            "name": "oem",
            "required": true,
            "schema": {
              "title": "Oem",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-device-id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "title": "X-Device-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "content-type",
            "required": false,
            "schema": {
              "default": "application/octet-stream",
              "maxLength": 80,
              "title": "Content-Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Send raw payloads of one source, newline separated for text formats",
        "tags": [
          "ingest"
        ]
      }
    },
...
    "/api/v1/oems": {
      "get": {
        "operationId": "oems_api_v1_oems_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Every telemetry source with its live mapping and health",
        "tags": [
          "registry"
        ]
      }
    },
...
    "/api/v1/schema/canonical": {
      "get": {
        "operationId": "schema_api_v1_schema_canonical_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "The canonical event: JSON Schema, field list, allowed transforms",
        "tags": [
          "registry"
        ]
      }
    },
...
  }
}